+++ app/controllers/api/v1/vt/vod_controller.rb
@@ -4,7 +4,7 @@ module Api
       class VodController < ApplicationController
         def index
           request_params = params.permit(Vehicle::Telematics::VodService::FETCH_QUERY_PARAMS)
-          service = Vehicle::Telematics::VodService.new(request_params)
+          service = ::Vt::VodService.new(request_params)
           response = service.fetch!
 
           response = if service.errors.present?
@@ -17,7 +17,7 @@ module Api
 
         def create
           request_params = params.permit(Vehicle::Telematics::VodService::CREATE_QUERY_PARAMS)
-          service = Vehicle::Telematics::VodService.new(request_params)
+          service = ::Vt::VodService.new(request_params)
           response = service.create!
           response = if service.errors.present?
                        Loconav::Response::Builder.failure(errors: response)
