+++ app/services/vt/livestream_service.rb
@@ -1,5 +1,5 @@
 module Vt
-  class VtLivestreamService
+  class LivestreamService
 
     include VtHelper
