+++ app/services/vt/vod_service.rb
